requirements.txt
altgraph==0.17
certifi==2019.6.16
chardet==3.0.4
Click==7.0
configobj==5.0.6
confusable-homoglyphs==3.2.0
cycler==0.10.0
defusedxml==0.6.0
Django==1.11
django-allauth==0.40.0
django-authorize==0.1.3
django-registration==3.0.1
django-reservation==0.2.9
django-widget-tweaks==1.4.8
djangorestframework==3.7.7
dnspython==1.16.0
fbs==0.8.6
Flask==1.1.1
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
macholib==1.14
MarkupSafe==1.1.1
matplotlib==3.2.1
mysql-connector==2.2.9
mysql-connector-python==8.0.19
numpy==1.18.2
oauthlib==3.1.0
pefile==2019.4.18
protobuf==3.6.1
psycopg2==2.8.4
psycopg2-binary==2.8.3
PyInstaller==3.4
pyparsing==2.4.6
PyQt5==5.13.0
PyQt5-sip==4.19.18
pyqt5-tools==5.13.0.1.5
python-dateutil==2.8.1
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.2
pywin32-ctypes==0.2.0
requests==2.22.0
requests-oauthlib==1.2.0
six==1.14.0
SQLAlchemy==1.3.8
sqlparse==0.3.0
urllib3==1.25.6
virtualenv==16.7.4
Werkzeug==0.16.0
wkhtmltopdf==0.2
xlrd==1.2.0
XlsxWriter==1.2.7
